politagree.ts created in the models directory
