+++ src/models/index.ts
@@ -13,6 +13,7 @@ export * from './MgmExceptionTypes'
 export * from './name-val'
 export * from './name-value'
 export * from './PagedResponse'
+export * from './politagree'
 export * from './ticker-info'
 export * from './types'
 export * from './UserState'
